2024/12/30
完成交付给正阳的所有功能

+++ main/elabel_state_machine/ElabelController.hpp
@@ -3,7 +3,10 @@
 
 #include "StateMachine.hpp"
 #include "global_message.h"
-#include "control_driver.h"
+#include "global_time.h"
+#include "global_draw.h"
+#include "control_driver.hpp"
+#include "ui.h"
 #define TimeCountdownOffset 300
 #define second_perencoder 150
 #define MAX_time 3600
@@ -31,14 +34,14 @@ class ElabelController
         void Init();
 
         //所选择的task的id(一定要以id来传递，不要用number来传递，id是唯一的)
-        uint8_t ChosenTaskId;
+        uint16_t ChosenTaskId;
         
         //任务倒计时的实际时间
         uint32_t TimeCountdown;
         //所选择的task在当前列表中的位置
-        uint8_t ChosenTaskNum;
+        uint16_t ChosenTaskNum;
         //当前task_length的长度
-        uint8_t TaskLength;
+        uint16_t TaskLength;
         //刷新标记
         bool need_flash_paper = false;
